+++ scripts/const.js
@@ -2,8 +2,12 @@ export const MODULE_ID = "objection";
 export const MS_TO_S = 1000;
 
 export const IMAGES = {
-  BENCH: "modules/objection/assets/lawyer_bench_basic_smol.webp",
+  BENCH: "modules/objection/assets/lawyer_Bench_Basic.webp",
   TEXT: {
-    OBJECTION: "modules/objection/assets/objection.webp",
+    objection: "modules/objection/assets/objection.webp",
+    eureka: "modules/objection/assets/eureka.webp",
+    gotcha: "modules/objection/assets/gotcha.webp",
+    hold_it: "modules/objection/assets/hold_it.webp",
+    take_that: "modules/objection/assets/take_that.webp",
   },
 };